refactor(http): change Service to Util Function approach and remove example code (service/component)
