--- src/components/atoms/example-request-component/styles.css
@@ -1,3 +1,0 @@
-.text {
-  color: var(--color-primary);
-}
