--- src/components/atoms/index.js
@@ -1,1 +1,0 @@
-export { default as ExampleRequestComponent } from './example-request-component'
