--- src/services/api/http-auth.js
@@ -1,19 +1,0 @@
-import axios from 'axios'
-
-import { requestAuthInterceptor } from './interceptors'
-
-const createAuthClient = () => {
-  const configs = {
-    baseURL: process.env.API_URL,
-  }
-
-  const client = axios.create(configs)
-
-  client.interceptors.request.use(requestAuthInterceptor, Promise.reject)
-
-  return client
-}
-
-const instance = createAuthClient()
-
-export default instance
