--- src/services/api/http.js
@@ -1,13 +1,0 @@
-import axios from 'axios'
-
-const createClient = () => {
-  const configs = {
-    baseURL: process.env.API_URL,
-  }
-
-  return axios.create(configs)
-}
-
-const instance = createClient()
-
-export default instance
