--- src/services/api/index.js
@@ -1,2 +1,0 @@
-export { default as http } from './http'
-export { default as httpAuth } from './http-auth'
