--- src/services/api/interceptors.js
@@ -1,12 +1,0 @@
-export const requestAuthInterceptor = config => {
-  const { headers: originalHeaders } = config
-  // place your auth token from your store here
-  const token = ''
-
-  const headers = {
-    ...originalHeaders,
-    Authorization: token,
-  }
-
-  return { ...config, headers }
-}
